Clean up some #defines...
git-svn-id: http://svn.osgeo.org/mapguide/trunk@2699 a442ad82-b40c-0410-ae44-e6d7bd6e4a36

+++ MgDev/Common/Foundation/System/Util.h
@@ -25,7 +25,7 @@
 
 #define DoubleMaxValue  DBL_MAX
 #define DoubleMinValue  DBL_MIN
-#define ROUND(x) (floor(x+0.5))
+#define ROUND(x) (int)(floor(x+0.5))
 
 #define DoubleNegativeInfinity numeric_limits<double>::infinity()
 #define DoublePositiveInfinity numeric_limits<double>::infinity()
